Several bug fixes
git-svn-id: https://svn.atnf.csiro.au/difx@3645 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/mk5daemon/trunk/src/mk5daemon.h
@@ -156,7 +156,7 @@ typedef struct
 	int driveStatsIndex[N_BANK];
 	double recordRate;	/* Mbps */
 	char dataSource[8];
-	unsigned long bitstreamMask;
+	int bitstreamMask;
 	int decimationRatio;
 	int unprotected;	/* set after protect=off */
